Stop using large umbrella modules for types
