+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Constraint/Split.hs
@@ -40,7 +40,16 @@ import           Language.Fixpoint.SortCheck (pruneUnsortedReft)
 
 import           Language.Haskell.Liquid.Misc -- (concatMapM)
 import qualified Language.Haskell.Liquid.UX.CTags       as Tg
-import           Language.Haskell.Liquid.Types hiding (loc)
+import           Language.Haskell.Liquid.Types.Errors
+import           Language.Haskell.Liquid.Types.Fresh
+import           Language.Haskell.Liquid.Types.PredType
+import           Language.Haskell.Liquid.Types.RefType
+import           Language.Haskell.Liquid.Types.RType
+import           Language.Haskell.Liquid.Types.RTypeOp
+import           Language.Haskell.Liquid.Types.Types
+import           Language.Haskell.Liquid.Types.Variance
+import           Language.Haskell.Liquid.UX.Config
+import           Language.Haskell.Liquid.UX.Tidy ()
 
 
 import           Language.Haskell.Liquid.Constraint.Types
