+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Constraint/Termination.hs
@@ -25,17 +25,18 @@ import           Control.Monad.State ( gets )
 import           Text.PrettyPrint.HughesPJ ( (<+>), text )
 import qualified Language.Haskell.Liquid.GHC.Misc                    as GM
 import qualified Language.Fixpoint.Types            as F
-import           Language.Fixpoint.Types.PrettyPrint (PPrint)
 import           Language.Haskell.Liquid.Constraint.Types (CG, CGInfo (..), CGEnv, makeRecInvariants)
 import           Language.Haskell.Liquid.Constraint.Monad (addWarning)
 import           Language.Haskell.Liquid.Constraint.Env (setTRec)
 import           Language.Haskell.Liquid.Constraint.Template ( Template(..), unTemplate, varTemplate, safeFromAsserted, extender )
-import           Language.Haskell.Liquid.Types (refreshArgs, HasConfig (..), toRSort)
+import           Language.Haskell.Liquid.Types.Errors
+import           Language.Haskell.Liquid.Types.Fresh
+import           Language.Haskell.Liquid.Types.RType
+import           Language.Haskell.Liquid.Types.RTypeOp
 import           Language.Haskell.Liquid.Types.Types
-  (SpecType, TError (..), RType (..), RTypeRep (..), Oblig (..), Error, Config (..), RReft,
-   toRTypeRep, structuralTerm, bkArrowDeep, mkArrow, bkUniv, bkArrow, fromRTypeRep)
 import           Language.Haskell.Liquid.Types.RefType (isDecreasing, makeDecrType, makeLexRefa, makeNumEnv)
 import           Language.Haskell.Liquid.Misc (safeFromLeft, replaceN, (<->), zip4, safeFromJust, fst5)
+import           Language.Haskell.Liquid.UX.Config
 import qualified Liquid.GHC.API as GHC
