+++ liquidhaskell-boot/src/Language/Haskell/Liquid/GHC/Interface.hs
@@ -79,7 +79,12 @@ import Language.Haskell.Liquid.WiredIn (isDerivedInstance)
 import qualified Language.Haskell.Liquid.Measure  as Ms
 import qualified Language.Haskell.Liquid.Misc     as Misc
 import Language.Haskell.Liquid.Parse
-import Language.Haskell.Liquid.Types hiding (Spec)
+import Language.Haskell.Liquid.Types.Errors
+import Language.Haskell.Liquid.Types.PrettyPrint
+import Language.Haskell.Liquid.Types.Specs
+import Language.Haskell.Liquid.Types.Types
+import Language.Haskell.Liquid.Types.Visitors
+import Language.Haskell.Liquid.UX.Config
 import Language.Haskell.Liquid.UX.Tidy
