+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Transforms/Simplify.hs
@@ -1,7 +1,7 @@
 module Language.Haskell.Liquid.Transforms.Simplify (simplifyBounds) where
 
 import Prelude hiding (error)
-import Language.Haskell.Liquid.Types
+import Language.Haskell.Liquid.Types.RType
 import Language.Fixpoint.Types
 import Language.Fixpoint.Types.Visitor
 -- import Control.Applicative                 ((<$>))
