+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Types/Equality.hs
@@ -5,7 +5,9 @@
 module Language.Haskell.Liquid.Types.Equality where
 
 import qualified Language.Fixpoint.Types as F
-import           Language.Haskell.Liquid.Types
+import           Language.Haskell.Liquid.Types.RType
+import           Language.Haskell.Liquid.Types.RefType ()
+import           Language.Haskell.Liquid.Types.Types
 import qualified Liquid.GHC.API as Ghc
 
 import Control.Monad (liftM2, zipWithM)
