+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Types/Fresh.hs
@@ -34,6 +34,9 @@ import           Prelude                        hiding (error)
 import qualified Language.Fixpoint.Types as F
 -- import           Language.Fixpoint.Types.Visitor (kvars)
 import           Language.Haskell.Liquid.Misc  (single)
+import           Language.Haskell.Liquid.Types.Errors
+import           Language.Haskell.Liquid.Types.RType
+import           Language.Haskell.Liquid.Types.RTypeOp
 import           Language.Haskell.Liquid.Types.Types
 import           Language.Haskell.Liquid.Types.RefType
