+++ liquidhaskell-boot/src/Language/Haskell/Liquid/Types/RefType.hs
@@ -111,7 +111,9 @@ import qualified Language.Fixpoint.Types as F
 import           Language.Haskell.Liquid.Types.Errors
 import           Language.Haskell.Liquid.Types.PrettyPrint
 
-import           Language.Haskell.Liquid.Types.Types hiding (R, DataConP (..))
+import           Language.Haskell.Liquid.Types.RType
+import           Language.Haskell.Liquid.Types.RTypeOp
+import           Language.Haskell.Liquid.Types.Types
 import           Language.Haskell.Liquid.Types.Variance
 import           Language.Haskell.Liquid.Misc
 import           Language.Haskell.Liquid.Types.Names
