+++ liquidhaskell-boot/src/Language/Haskell/Liquid/WiredIn.hs
@@ -32,6 +32,8 @@ import Prelude                                hiding (error)
 import Language.Haskell.Liquid.GHC.Misc
 import qualified Liquid.GHC.API as Ghc
 import Liquid.GHC.API (Var, Arity, TyVar, Bind(..), Boxity(..), Expr(..), ForAllTyFlag(Required))
+import Language.Haskell.Liquid.Types.Errors
+import Language.Haskell.Liquid.Types.RType
 import Language.Haskell.Liquid.Types.Types
 import Language.Haskell.Liquid.Types.RefType
 import Language.Haskell.Liquid.Types.Variance
